Modified interface by adding &opts and moved opts.
Neovim passes arguments in a slightly weird way, because it also has to
pass special options user specifies he wants such as range, register,
bang, etc. . To achieve that it wraps all the actual arguments into a
list, which gets sent off as a first argument, and the rest of options
get sent in a specified order as the remaining arguments.

This commit adds a special lambda list specifier &opts, which should be
added to the end of lambda list and is used to get neovim's options. For
a better description you should refer to README.

Commit also now expects (opts [plist]) inside a declare expression, with
plist converted to alist to better mimic other declare expressions
(refer to sample plugin to see an example).